#88: Preparations for implementing file manager imitation
Application now tracks if it launches for the first time or not.
If yes, the Core Data managed database is initialized with "Encrypted" and "Decrypted" folders instances.
Library View Controller created, which is supposed to imitate file manager. For now it just displays all the folders from the database — lotta stuff to be implemented.

Unit tests and UI tests to be added.